adjusted namespace of getter function for state
